requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
websockets==12.0
pydantic==2.9.0
pydantic-settings==2.5.0
langgraph==0.2.0
google-generativeai==0.8.0
snowflake-connector-python==3.12.0
ibm-watsonx-ai==1.1.0
structlog==24.4.0
pyyaml==6.0.2
pytest==8.3.0
httpx==0.27.0
